Split out parameter classes into their own files
