+++ libraries/ts-command-line/src/test/ActionlessParser.test.ts
@@ -1,9 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
 // See LICENSE in the project root for license information.
 
-import { CommandLineParser } from '../providers/CommandLineParser';
-import { CommandLineFlagParameter } from '../parameters/CommandLineParameter';
-
+import { CommandLineParser, CommandLineFlagParameter } from '..';
 
 class TestCommandLine extends CommandLineParser {
   public flag: CommandLineFlagParameter;
