+++ libraries/ts-command-line/src/test/CommandLineParameter.test.ts
@@ -3,10 +3,7 @@
 
 import * as colors from 'colors';
 
-import { CommandLineAction } from '../providers/CommandLineAction';
-import { CommandLineParser } from '../providers/CommandLineParser';
-import { DynamicCommandLineParser } from '../providers/DynamicCommandLineParser';
-import { DynamicCommandLineAction } from '../providers/DynamicCommandLineAction';
+import { CommandLineAction, CommandLineParser, DynamicCommandLineParser, DynamicCommandLineAction } from '..';
 
 function createParser(): DynamicCommandLineParser {
   const commandLineParser: DynamicCommandLineParser = new DynamicCommandLineParser(
