+++ libraries/ts-command-line/src/test/DynamicCommandLineParser.test.ts
@@ -1,9 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
 // See LICENSE in the project root for license information.
 
-import { DynamicCommandLineParser } from '../providers/DynamicCommandLineParser';
-import { DynamicCommandLineAction } from '../providers/DynamicCommandLineAction';
-import { CommandLineFlagParameter } from '../parameters/CommandLineParameter';
+import { DynamicCommandLineParser, DynamicCommandLineAction, CommandLineFlagParameter } from '..';
 
 describe('DynamicCommandLineParser', () => {
